tsla_forecast_app/requirements_trading.txt
# Tesla Trading Robot - Advanced Trading System Requirements
# ========================================================

# Core PyQt5
PyQt5==5.15.9

# Environment variables
python-dotenv==1.0.0

# HTTP requests for real-time data
requests==2.31.0

# Data processing and analysis
pandas==2.0.3
numpy==1.24.3

# Technical analysis indicators
TA-Lib==0.4.28

# Machine learning for advanced strategies
scikit-learn==1.3.0

# Financial data
yfinance==0.2.18

# Database for trade history
sqlite3

# Charting and visualization
matplotlib==3.7.2
plotly==5.15.0

# Risk management
scipy==1.11.1

# Development and testing
pytest==7.4.0
pytest-qt==4.2.0

# Optional: Advanced features
# alpaca-trade-api==3.0.0  # For live trading
# ibapi==10.19.1           # Interactive Brokers API
# ccxt==4.0.0              # Cryptocurrency exchanges
